+++ .eslintrc.js
@@ -1,0 +1,11 @@
+module.exports = {
+    "extends": "eslint-config-airbnb",
+    "rules": {
+        // enable additional rules
+        "indent": ["error", 2],
+        "linebreak-style": ["error", "unix"],
+
+        "no-underscore-dangle": "off",
+        "no-continue": "off",
+    },
+};
