+++ package.json
@@ -1,22 +1,46 @@
 {
   "name": "json-stream-stringify",
-  "version": "1.6.1",
+  "version": "2.0.0",
   "description": "JSON.Stringify as a readable stream",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/Faleij/json-stream-stringify.git"
   },
-  "main": "jsonStreamify.js",
-  "dependencies": {},
+  "browser": "dist/es5.umd.min.js",
+  "main": "dist/es6.umd.min.js",
+  "module": "src/JsonStreamify.js",
+  "peerDependencies": {
+    "babel-runtime": "*",
+    "babel-polyfill": "*"
+  },
   "devDependencies": {
+    "babel-core": "6.26.3",
+    "babel-plugin-transform-runtime": "6.23.0",
+    "babel-polyfill": "^6.26.0",
+    "babel-preset-env": "1.7.0",
+    "babel-runtime": "^6.26.0",
     "coveralls": "3.0.1",
-    "istanbul": "0.4.5",
+    "eslint": "^4.9.0",
+    "eslint-config-airbnb": "^16.1.0",
+    "eslint-plugin-import": "^2.7.0",
+    "eslint-plugin-jsx-a11y": "^6.0.2",
+    "eslint-plugin-mocha": "^5.0.0",
+    "eslint-plugin-react": "^7.4.0",
     "expect.js": "0.3.1",
-    "mocha": "5.2.0"
+    "istanbul": "0.4.5",
+    "mocha": "5.2.0",
+    "nyc": "^12.0.2",
+    "rollup": "0.60.4",
+    "rollup-plugin-babel": "3.0.4",
+    "rollup-plugin-node-globals": "^1.2.1",
+    "rollup-plugin-terser": "1.0.1",
+    "rollup-plugin-uglify": "4.0.0"
   },
   "scripts": {
-    "test": "mocha",
-    "coverage": "istanbul cover node_modules/mocha/bin/_mocha -- test.js -R spec"
+    "build": "rollup -c rollup.config.js",
+    "test": "mocha --require babel-core/register -R spec",
+    "coverage": "nyc npm test",
+    "coveralls": "npm run coverage && nyc report --reporter=text-lcov | coveralls"
   },
   "keywords": [
     "JSON",
