+++ rollup.es5.entry.js
@@ -1,0 +1,3 @@
+import "babel-polyfill";
+
+export { default } from "./src/JsonStreamify";
