+++ rollup.es6.entry.js
@@ -1,0 +1,1 @@
+export { default } from "./src/JsonStreamify";
