+++ test/.babelrc
@@ -1,0 +1,9 @@
+{
+    "presets": [
+        ["env", {
+            "targets": {
+                "node": "current"
+            }
+        }]
+    ]
+}
