+++ test/.eslintrc.js
@@ -1,0 +1,12 @@
+module.exports = {
+    extends: ['../.eslintrc.js'],
+    plugins: [
+        'mocha',
+    ],
+    env: {
+        mocha: true,
+    },
+    rules: {
+        'no-sparse-arrays': 'off',
+    },
+};
